requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
boto3==1.38.36
botocore==1.38.36
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.11.3
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
fonttools==4.58.4
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.11
numpy==2.3.0
openai==1.86.0
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.63
